Refactoring to odometer.cpp/h, added gps_bearing.
